refactoring: split up asn.rs into smaller pieces (ext of types)
